Refactor headers in panda/board/stm32h7/*.h
