--- board/stm32h7/llusb_declarations.h
@@ -1,16 +1,0 @@
-#pragma once
-
-extern USB_OTG_GlobalTypeDef *USBx;
-
-#define USBx_DEVICE     ((USB_OTG_DeviceTypeDef *)((uint32_t)USBx + USB_OTG_DEVICE_BASE))
-#define USBx_INEP(i)    ((USB_OTG_INEndpointTypeDef *)((uint32_t)USBx + USB_OTG_IN_ENDPOINT_BASE + ((i) * USB_OTG_EP_REG_SIZE)))
-#define USBx_OUTEP(i)   ((USB_OTG_OUTEndpointTypeDef *)((uint32_t)USBx + USB_OTG_OUT_ENDPOINT_BASE + ((i) * USB_OTG_EP_REG_SIZE)))
-#define USBx_DFIFO(i)   *(__IO uint32_t *)((uint32_t)USBx + USB_OTG_FIFO_BASE + ((i) * USB_OTG_FIFO_SIZE))
-#define USBx_PCGCCTL    *(__IO uint32_t *)((uint32_t)USBx + USB_OTG_PCGCCTL_BASE)
-
-#define USBD_FS_TRDT_VALUE        6UL
-#define USB_OTG_SPEED_FULL        3U
-#define DCFG_FRAME_INTERVAL_80    0U
-
-void usb_irqhandler(void);
-void usb_init(void);
