+++ board/stm32h7/src/interrupt_handlers.c
@@ -1,0 +1,144 @@
+#include "../interrupt_handlers.h"
+// ********************* Bare interrupt handlers *********************
+// Interrupts for STM32H7x5
+
+void WWDG_IRQHandler(void) {handle_interrupt(WWDG_IRQn);}
+void PVD_AVD_IRQHandler(void) {handle_interrupt(PVD_AVD_IRQn);}
+void TAMP_STAMP_IRQHandler(void) {handle_interrupt(TAMP_STAMP_IRQn);}
+void RTC_WKUP_IRQHandler(void) {handle_interrupt(RTC_WKUP_IRQn);}
+void FLASH_IRQHandler(void) {handle_interrupt(FLASH_IRQn);}
+void RCC_IRQHandler(void) {handle_interrupt(RCC_IRQn);}
+void EXTI0_IRQHandler(void) {handle_interrupt(EXTI0_IRQn);}
+void EXTI1_IRQHandler(void) {handle_interrupt(EXTI1_IRQn);}
+void EXTI2_IRQHandler(void) {handle_interrupt(EXTI2_IRQn);}
+void EXTI3_IRQHandler(void) {handle_interrupt(EXTI3_IRQn);}
+void EXTI4_IRQHandler(void) {handle_interrupt(EXTI4_IRQn);}
+void DMA1_Stream0_IRQHandler(void) {handle_interrupt(DMA1_Stream0_IRQn);}
+void DMA1_Stream1_IRQHandler(void) {handle_interrupt(DMA1_Stream1_IRQn);}
+void DMA1_Stream2_IRQHandler(void) {handle_interrupt(DMA1_Stream2_IRQn);}
+void DMA1_Stream3_IRQHandler(void) {handle_interrupt(DMA1_Stream3_IRQn);}
+void DMA1_Stream4_IRQHandler(void) {handle_interrupt(DMA1_Stream4_IRQn);}
+void DMA1_Stream5_IRQHandler(void) {handle_interrupt(DMA1_Stream5_IRQn);}
+void DMA1_Stream6_IRQHandler(void) {handle_interrupt(DMA1_Stream6_IRQn);}
+void ADC_IRQHandler(void) {handle_interrupt(ADC_IRQn);}
+void FDCAN1_IT0_IRQHandler(void) {handle_interrupt(FDCAN1_IT0_IRQn);}
+void FDCAN2_IT0_IRQHandler(void) {handle_interrupt(FDCAN2_IT0_IRQn);}
+void FDCAN1_IT1_IRQHandler(void) {handle_interrupt(FDCAN1_IT1_IRQn);}
+void FDCAN2_IT1_IRQHandler(void) {handle_interrupt(FDCAN2_IT1_IRQn);}
+void EXTI9_5_IRQHandler(void) {handle_interrupt(EXTI9_5_IRQn);}
+void TIM1_BRK_IRQHandler(void) {handle_interrupt(TIM1_BRK_IRQn);}
+void TIM1_UP_TIM10_IRQHandler(void) {handle_interrupt(TIM1_UP_TIM10_IRQn);}
+void TIM1_TRG_COM_IRQHandler(void) {handle_interrupt(TIM1_TRG_COM_IRQn);}
+void TIM1_CC_IRQHandler(void) {handle_interrupt(TIM1_CC_IRQn);}
+void TIM2_IRQHandler(void) {handle_interrupt(TIM2_IRQn);}
+void TIM3_IRQHandler(void) {handle_interrupt(TIM3_IRQn);}
+void TIM4_IRQHandler(void) {handle_interrupt(TIM4_IRQn);}
+void I2C1_EV_IRQHandler(void) {handle_interrupt(I2C1_EV_IRQn);}
+void I2C1_ER_IRQHandler(void) {handle_interrupt(I2C1_ER_IRQn);}
+void I2C2_EV_IRQHandler(void) {handle_interrupt(I2C2_EV_IRQn);}
+void I2C2_ER_IRQHandler(void) {handle_interrupt(I2C2_ER_IRQn);}
+void SPI1_IRQHandler(void) {handle_interrupt(SPI1_IRQn);}
+void SPI2_IRQHandler(void) {handle_interrupt(SPI2_IRQn);}
+void USART1_IRQHandler(void) {handle_interrupt(USART1_IRQn);}
+void USART2_IRQHandler(void) {handle_interrupt(USART2_IRQn);}
+void USART3_IRQHandler(void) {handle_interrupt(USART3_IRQn);}
+void EXTI15_10_IRQHandler(void) {handle_interrupt(EXTI15_10_IRQn);}
+void RTC_Alarm_IRQHandler(void) {handle_interrupt(RTC_Alarm_IRQn);}
+void TIM8_BRK_TIM12_IRQHandler(void) {handle_interrupt(TIM8_BRK_TIM12_IRQn);}
+void TIM8_UP_TIM13_IRQHandler(void) {handle_interrupt(TIM8_UP_TIM13_IRQn);}
+void TIM8_TRG_COM_TIM14_IRQHandler(void) {handle_interrupt(TIM8_TRG_COM_TIM14_IRQn);}
+void TIM8_CC_IRQHandler(void) {handle_interrupt(TIM8_CC_IRQn);}
+void DMA1_Stream7_IRQHandler(void) {handle_interrupt(DMA1_Stream7_IRQn);}
+void FMC_IRQHandler(void) {handle_interrupt(FMC_IRQn);}
+void SDMMC1_IRQHandler(void) {handle_interrupt(SDMMC1_IRQn);}
+void TIM5_IRQHandler(void) {handle_interrupt(TIM5_IRQn);}
+void SPI3_IRQHandler(void) {handle_interrupt(SPI3_IRQn);}
+void UART4_IRQHandler(void) {handle_interrupt(UART4_IRQn);}
+void UART5_IRQHandler(void) {handle_interrupt(UART5_IRQn);}
+void TIM6_DAC_IRQHandler(void) {handle_interrupt(TIM6_DAC_IRQn);}
+void TIM7_IRQHandler(void) {handle_interrupt(TIM7_IRQn);}
+void DMA2_Stream0_IRQHandler(void) {handle_interrupt(DMA2_Stream0_IRQn);}
+void DMA2_Stream1_IRQHandler(void) {handle_interrupt(DMA2_Stream1_IRQn);}
+void DMA2_Stream2_IRQHandler(void) {handle_interrupt(DMA2_Stream2_IRQn);}
+void DMA2_Stream3_IRQHandler(void) {handle_interrupt(DMA2_Stream3_IRQn);}
+void DMA2_Stream4_IRQHandler(void) {handle_interrupt(DMA2_Stream4_IRQn);}
+void ETH_IRQHandler(void) {handle_interrupt(ETH_IRQn);}
+void ETH_WKUP_IRQHandler(void) {handle_interrupt(ETH_WKUP_IRQn);}
+void FDCAN_CAL_IRQHandler(void) {handle_interrupt(FDCAN_CAL_IRQn);}
+void DMA2_Stream5_IRQHandler(void) {handle_interrupt(DMA2_Stream5_IRQn);}
+void DMA2_Stream6_IRQHandler(void) {handle_interrupt(DMA2_Stream6_IRQn);}
+void DMA2_Stream7_IRQHandler(void) {handle_interrupt(DMA2_Stream7_IRQn);}
+void USART6_IRQHandler(void) {handle_interrupt(USART6_IRQn);}
+void I2C3_EV_IRQHandler(void) {handle_interrupt(I2C3_EV_IRQn);}
+void I2C3_ER_IRQHandler(void) {handle_interrupt(I2C3_ER_IRQn);}
+void OTG_HS_EP1_OUT_IRQHandler(void) {handle_interrupt(OTG_HS_EP1_OUT_IRQn);}
+void OTG_HS_EP1_IN_IRQHandler(void) {handle_interrupt(OTG_HS_EP1_IN_IRQn);}
+void OTG_HS_WKUP_IRQHandler(void) {handle_interrupt(OTG_HS_WKUP_IRQn);}
+void OTG_HS_IRQHandler(void) {handle_interrupt(OTG_HS_IRQn);}
+void DCMI_PSSI_IRQHandler(void) {handle_interrupt(DCMI_PSSI_IRQn);}
+void CRYP_IRQHandler(void) {handle_interrupt(CRYP_IRQn);}
+void HASH_RNG_IRQHandler(void) {handle_interrupt(HASH_RNG_IRQn);}
+void FPU_IRQHandler(void) {handle_interrupt(FPU_IRQn);}
+void UART7_IRQHandler(void) {handle_interrupt(UART7_IRQn);}
+void UART8_IRQHandler(void) {handle_interrupt(UART8_IRQn);}
+void SPI4_IRQHandler(void) {handle_interrupt(SPI4_IRQn);}
+void SPI5_IRQHandler(void) {handle_interrupt(SPI5_IRQn);}
+void SPI6_IRQHandler(void) {handle_interrupt(SPI6_IRQn);}
+void SAI1_IRQHandler(void) {handle_interrupt(SAI1_IRQn);}
+void LTDC_IRQHandler(void) {handle_interrupt(LTDC_IRQn);}
+void LTDC_ER_IRQHandler(void) {handle_interrupt(LTDC_ER_IRQn);}
+void DMA2D_IRQHandler(void) {handle_interrupt(DMA2D_IRQn);}
+void OCTOSPI1_IRQHandler(void) {handle_interrupt(OCTOSPI1_IRQn);}
+void LPTIM1_IRQHandler(void) {handle_interrupt(LPTIM1_IRQn);}
+void CEC_IRQHandler(void) {handle_interrupt(CEC_IRQn);}
+void I2C4_EV_IRQHandler(void) {handle_interrupt(I2C4_EV_IRQn);}
+void I2C4_ER_IRQHandler(void) {handle_interrupt(I2C4_ER_IRQn);}
+void SPDIF_RX_IRQHandler(void) {handle_interrupt(SPDIF_RX_IRQn);}
+void DMAMUX1_OVR_IRQHandler(void) {handle_interrupt(DMAMUX1_OVR_IRQn);}
+void DFSDM1_FLT0_IRQHandler(void) {handle_interrupt(DFSDM1_FLT0_IRQn);}
+void DFSDM1_FLT1_IRQHandler(void) {handle_interrupt(DFSDM1_FLT1_IRQn);}
+void DFSDM1_FLT2_IRQHandler(void) {handle_interrupt(DFSDM1_FLT2_IRQn);}
+void DFSDM1_FLT3_IRQHandler(void) {handle_interrupt(DFSDM1_FLT3_IRQn);}
+void SWPMI1_IRQHandler(void) {handle_interrupt(SWPMI1_IRQn);}
+void TIM15_IRQHandler(void) {handle_interrupt(TIM15_IRQn);}
+void TIM16_IRQHandler(void) {handle_interrupt(TIM16_IRQn);}
+void TIM17_IRQHandler(void) {handle_interrupt(TIM17_IRQn);}
+void MDIOS_WKUP_IRQHandler(void) {handle_interrupt(MDIOS_WKUP_IRQn);}
+void MDIOS_IRQHandler(void) {handle_interrupt(MDIOS_IRQn);}
+void MDMA_IRQHandler(void) {handle_interrupt(MDMA_IRQn);}
+void SDMMC2_IRQHandler(void) {handle_interrupt(SDMMC2_IRQn);}
+void HSEM1_IRQHandler(void) {handle_interrupt(HSEM1_IRQn);}
+void ADC3_IRQHandler(void) {handle_interrupt(ADC3_IRQn);}
+void DMAMUX2_OVR_IRQHandler(void) {handle_interrupt(DMAMUX2_OVR_IRQn);}
+void BDMA_Channel0_IRQHandler(void) {handle_interrupt(BDMA_Channel0_IRQn);}
+void BDMA_Channel1_IRQHandler(void) {handle_interrupt(BDMA_Channel1_IRQn);}
+void BDMA_Channel2_IRQHandler(void) {handle_interrupt(BDMA_Channel2_IRQn);}
+void BDMA_Channel3_IRQHandler(void) {handle_interrupt(BDMA_Channel3_IRQn);}
+void BDMA_Channel4_IRQHandler(void) {handle_interrupt(BDMA_Channel4_IRQn);}
+void BDMA_Channel5_IRQHandler(void) {handle_interrupt(BDMA_Channel5_IRQn);}
+void BDMA_Channel6_IRQHandler(void) {handle_interrupt(BDMA_Channel6_IRQn);}
+void BDMA_Channel7_IRQHandler(void) {handle_interrupt(BDMA_Channel7_IRQn);}
+void COMP_IRQHandler(void) {handle_interrupt(COMP_IRQn);}
+void LPTIM2_IRQHandler(void) {handle_interrupt(LPTIM2_IRQn);}
+void LPTIM3_IRQHandler(void) {handle_interrupt(LPTIM3_IRQn);}
+void LPTIM4_IRQHandler(void) {handle_interrupt(LPTIM4_IRQn);}
+void LPTIM5_IRQHandler(void) {handle_interrupt(LPTIM5_IRQn);}
+void LPUART1_IRQHandler(void) {handle_interrupt(LPUART1_IRQn);}
+void CRS_IRQHandler(void) {handle_interrupt(CRS_IRQn);}
+void ECC_IRQHandler(void) {handle_interrupt(ECC_IRQn);}
+void SAI4_IRQHandler(void) {handle_interrupt(SAI4_IRQn);}
+void DTS_IRQHandler(void) {handle_interrupt(DTS_IRQn);}
+void WAKEUP_PIN_IRQHandler(void) {handle_interrupt(WAKEUP_PIN_IRQn);}
+void OCTOSPI2_IRQHandler(void) {handle_interrupt(OCTOSPI2_IRQn);}
+void OTFDEC1_IRQHandler(void) {handle_interrupt(OTFDEC1_IRQn);}
+void OTFDEC2_IRQHandler(void) {handle_interrupt(OTFDEC2_IRQn);}
+void FMAC_IRQHandler(void) {handle_interrupt(FMAC_IRQn);}
+void CORDIC_IRQHandler(void) {handle_interrupt(CORDIC_IRQn);}
+void UART9_IRQHandler(void) {handle_interrupt(UART9_IRQn);}
+void USART10_IRQHandler(void) {handle_interrupt(USART10_IRQn);}
+void I2C5_EV_IRQHandler(void) {handle_interrupt(I2C5_EV_IRQn);}
+void I2C5_ER_IRQHandler(void) {handle_interrupt(I2C5_ER_IRQn);}
+void FDCAN3_IT0_IRQHandler(void) {handle_interrupt(FDCAN3_IT0_IRQn);}
+void FDCAN3_IT1_IRQHandler(void) {handle_interrupt(FDCAN3_IT1_IRQn);}
+void TIM23_IRQHandler(void) {handle_interrupt(TIM23_IRQn);}
+void TIM24_IRQHandler(void) {handle_interrupt(TIM24_IRQn);}
